Uploaded recommendationAlgo.py, dbPopulate.py, recommendations.text and editted db.sql.
db.sql changes - added column for name and filled with random samples, changed star locations so 1 - 50 is filled and none are in the same location twice.

recommendationAlgo.py - creates recommendations.text when ran, for optimum results this should be called and then the text pulled regularly.

recommendations.text - 10 recommendations for highest rev slots in casino for website to pull

dbPopulate - creates the slot_machine.db file